requirements.txt
streamlit==1.2.0
Pillow==8.4.0